feat: EIP-6372-compliant `voteStart` and `voteEnd`
- also reduce use of `PureEpochs.currentEpoch()` and use `clock()`

+++ src/interfaces/IStandardGovernor.sol
@@ -124,8 +124,8 @@ interface IStandardGovernor is IBatchGovernor {
         external
         view
         returns (
-            uint16 voteStart,
-            uint16 voteEnd,
+            uint48 voteStart,
+            uint48 voteEnd,
             bool executed,
             ProposalState state,
             uint256 noVotes,
